+++ src/components/sections/cta-section.tsx
@@ -1,0 +1,56 @@
+'use client'
+
+import { motion } from 'framer-motion'
+import { WalletConnect } from '@/components/wallet-connect'
+
+export const CTASection = () => {
+  return (
+    <div className="bg-white">
+      <div className="mx-auto max-w-7xl py-24 sm:px-6 sm:py-32 lg:px-8">
+        <div className="relative isolate overflow-hidden bg-gradient-to-br from-indigo-900 via-indigo-800 to-purple-900 px-6 py-24 text-center shadow-2xl sm:rounded-3xl sm:px-16">
+          <motion.div
+            initial={{ opacity: 0, y: 20 }}
+            whileInView={{ opacity: 1, y: 0 }}
+            viewport={{ once: true }}
+            transition={{ duration: 0.5 }}
+          >
+            <h2 className="mx-auto max-w-2xl text-3xl font-bold tracking-tight text-white sm:text-4xl">
+              Start building your credit today
+            </h2>
+            <p className="mx-auto mt-6 max-w-xl text-lg leading-8 text-gray-300">
+              Connect your wallet and start using Kori to build your credit score while enjoying instant merchant payments.
+            </p>
+          </motion.div>
+          <div className="mt-10 flex items-center justify-center gap-x-6">
+            <WalletConnect />
+            <a
+              href="#"
+              className="text-sm font-semibell leading-6 text-white"
+            >
+              Learn more <span aria-hidden="true">→</span>
+            </a>
+          </div>
+          <svg
+            viewBox="0 0 1024 1024"
+            className="absolute left-1/2 top-1/2 -z-10 h-[64rem] w-[64rem] -translate-x-1/2 [mask-image:radial-gradient(closest-side,white,transparent)]"
+            aria-hidden="true"
+          >
+            <circle
+              cx={512}
+              cy={512}
+              r={512}
+              fill="url(#indigo-gradient)"
+              fillOpacity="0.3"
+            />
+            <defs>
+              <radialGradient id="indigo-gradient">
+                <stop stopColor="#6366f1" />
+                <stop offset={1} stopColor="#8b5cf6" />
+              </radialGradient>
+            </defs>
+          </svg>
+        </div>
+      </div>
+    </div>
+  )
+}
