+++ src/components/sections/index.ts
@@ -1,0 +1,6 @@
+export { HeroSection } from './hero-section'
+export { FeaturesSection } from './features-section'
+export { TierSection } from './tier-section'
+export { StatsSection } from './stats-section'
+export { HowItWorksSection } from './how-it-works-section'
+export { CTASection } from './cta-section'
